+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] - 2023-06-01
+- Add client and template for creating multiple stock equity baskets (no corax)
+- Add client and template for creating a single stock total return index with corax
+
 ## [0.9.0] - 2023-05-25
 - Make base classes private now that nothing is importing them directly.